Fix unit tests after audit log event types refactoring
